fix update_images check TINYURL

+++ scripts/EVE-NG/update_images.sh
@@ -23,27 +23,27 @@ do
         FILE=`echo $line | awk -F\| '{ print $2 }'`
         SHA1=`echo $line | awk -F\| '{ print $3 }'`
         TINYURL=`echo $line | awk -F\| '{ print $4 }'`
-        URL=`curl -s $TINYURL | cut -b 23-`
-
-        FileLocal=${BASE}${DIR}${FILE}
-
-        if test -d ${BASE}${DIR} ; then
-                if test -f ${FileLocal} ; then
-                        sha1=`sha1sum ${FileLocal} | awk '{ print $1}' `
-                        if [ $SHA1 != $sha1 ]; then
-                                rm -rf ${FileLocal}
+        if ! test -z ${TINYURL} ; then        
+                URL=`curl -s $TINYURL | cut -b 23-`
+                FileLocal=${BASE}${DIR}${FILE}
+                if test -d ${BASE}${DIR} ; then
+                        if test -f ${FileLocal} ; then
+                                sha1=`sha1sum ${FileLocal} | awk '{ print $1}' `
+                                if [ $SHA1 != $sha1 ]; then
+                                        rm -rf ${FileLocal}
+                                        echo "Downloading ${FILE} in ${BASE}${DIR}..." 
+                                        mkdir -p ${BASE}${DIR} && megadl --path=${FileLocal} ${URL} 
+        	                        else
+                                        echo "${FileLocal} is update" 
+                                fi
+                        else
                                 echo "Downloading ${FILE} in ${BASE}${DIR}..." 
                                 mkdir -p ${BASE}${DIR} && megadl --path=${FileLocal} ${URL} 
-	                        else
-                                echo "${FileLocal} is update" 
                         fi
                 else
                         echo "Downloading ${FILE} in ${BASE}${DIR}..." 
                         mkdir -p ${BASE}${DIR} && megadl --path=${FileLocal} ${URL} 
                 fi
-        else
-                echo "Downloading ${FILE} in ${BASE}${DIR}..." 
-                mkdir -p ${BASE}${DIR} && megadl --path=${FileLocal} ${URL} 
         fi
 done < <( cat ${BASE}/${LIST} )
